Switch navTgtPos BB to int32_t

+++ src/main/navigation/navigation.h
@@ -413,7 +413,7 @@ extern navSystemStatus_t    NAV_Status;
 extern int16_t navCurrentState;
 extern int16_t navActualVelocity[3];
 extern int16_t navDesiredVelocity[3];
-extern int16_t navTargetPosition[3];
+extern int32_t navTargetPosition[3];
 extern int32_t navLatestActualPosition[3];
 extern int16_t navActualSurface;
 extern uint16_t navFlags;
